chatbot/requirements.txt

pypdf
plotly
pyttsx3
prophet
textblob
chromadb
openpyxl
docx2txt
networkx
streamlit
ipykernel
setuptools
python-pptx
unstructured
google-genai
python-dotenv
langchain-groq
msoffcrypto-tool
langchain-openai
langchain-chroma
langchain-ollama
speechrecognition
langchain-community
google-generativeai
langchain-google-genai
streamlit-authenticator
openai
langchain_openai
langchain
crewai
crewai_tools

-e .